Several bug fixes, some related to memory issues.

+++ src/core/monitors/ScreenMonitor.cpp
@@ -101,10 +101,10 @@ void ScreenMonitor::monitor(unsigned long gen)
         std::stringstream ss;
         
         // set column width
-        size_t columnWidth = 12;
+        int columnWidth = 12;
         
         // set cycle column width
-        size_t cycleWidth = floor( log10( numCycles ) ) + 1;
+        int cycleWidth = floor( log10( numCycles ) ) + 1;
         cycleWidth = 5 > cycleWidth ? 5 : cycleWidth;
 
         // print the cycle number
@@ -222,9 +222,9 @@ void ScreenMonitor::printHeader( void )
     // print one column for the iteration number
     std::string header = "Iter";
 
-    size_t width = 4;
+    int width = 4;
     
-    size_t numWidth = size_t( log10( numCycles ) ) + 1;
+    int numWidth = int( log10( numCycles ) ) + 1;
     width = width > numWidth ? width : numWidth;
     
     size_t columnWidth = 12;
